Split fixtures in the Permission module
