+++ saleor/permission/tests/fixtures/permission.py
@@ -1,0 +1,128 @@
+import pytest
+
+from ...models import Permission
+
+
+@pytest.fixture
+def permission_manage_discounts():
+    return Permission.objects.get(codename="manage_discounts")
+
+
+@pytest.fixture
+def permission_manage_gift_card():
+    return Permission.objects.get(codename="manage_gift_card")
+
+
+@pytest.fixture
+def permission_manage_orders():
+    return Permission.objects.get(codename="manage_orders")
+
+
+@pytest.fixture
+def permission_manage_orders_import():
+    return Permission.objects.get(codename="manage_orders_import")
+
+
+@pytest.fixture
+def permission_manage_checkouts():
+    return Permission.objects.get(codename="manage_checkouts")
+
+
+@pytest.fixture
+def permission_handle_checkouts():
+    return Permission.objects.get(codename="handle_checkouts")
+
+
+@pytest.fixture
+def permission_manage_plugins():
+    return Permission.objects.get(codename="manage_plugins")
+
+
+@pytest.fixture
+def permission_manage_apps():
+    return Permission.objects.get(codename="manage_apps")
+
+
+@pytest.fixture
+def permission_handle_taxes():
+    return Permission.objects.get(codename="handle_taxes")
+
+
+@pytest.fixture
+def permission_manage_observability():
+    return Permission.objects.get(codename="manage_observability")
+
+
+@pytest.fixture
+def permission_manage_taxes():
+    return Permission.objects.get(codename="manage_taxes")
+
+
+@pytest.fixture
+def permission_manage_staff():
+    return Permission.objects.get(codename="manage_staff")
+
+
+@pytest.fixture
+def permission_manage_products():
+    return Permission.objects.get(codename="manage_products")
+
+
+@pytest.fixture
+def permission_manage_product_types_and_attributes():
+    return Permission.objects.get(codename="manage_product_types_and_attributes")
+
+
+@pytest.fixture
+def permission_manage_shipping():
+    return Permission.objects.get(codename="manage_shipping")
+
+
+@pytest.fixture
+def permission_manage_users():
+    return Permission.objects.get(codename="manage_users")
+
+
+@pytest.fixture
+def permission_impersonate_user():
+    return Permission.objects.get(codename="impersonate_user")
+
+
+@pytest.fixture
+def permission_manage_settings():
+    return Permission.objects.get(codename="manage_settings")
+
+
+@pytest.fixture
+def permission_manage_menus():
+    return Permission.objects.get(codename="manage_menus")
+
+
+@pytest.fixture
+def permission_manage_pages():
+    return Permission.objects.get(codename="manage_pages")
+
+
+@pytest.fixture
+def permission_manage_page_types_and_attributes():
+    return Permission.objects.get(codename="manage_page_types_and_attributes")
+
+
+@pytest.fixture
+def permission_manage_translations():
+    return Permission.objects.get(codename="manage_translations")
+
+
+@pytest.fixture
+def permission_manage_webhooks():
+    return Permission.objects.get(codename="manage_webhooks")
+
+
+@pytest.fixture
+def permission_manage_channels():
+    return Permission.objects.get(codename="manage_channels")
+
+
+@pytest.fixture
+def permission_manage_payments():
+    return Permission.objects.get(codename="handle_payments")
